refactor(dsh-provider-usage): aggregator 纯函数拆分——压实转换与查询投影独立成模块（阶段三 #670）
改动摘要：
- 新增 src/trend/aggregate-rows.ts：压实转换纯函数——emptyAggRow/emptyDirRow/
  emptyHourRow、addDetailTo/addCounterTo、mergeCell、mergeAggRows/mergeDirRows/
  mergeHourRows、emptyCell、sub/diffToken（token null-aware 差，校正回退用）
- 新增 src/trend/aggregate-query.ts：查询投影纯函数与类型——metricValue、
  周/月序列与区间工具（weekStartKey/lastNWeekKeys/lastNMonthKeys/monthRange/
  weekRange）、buildDirRows（#633 目录面：dirDays 单源快照 + 每日残差归未识别）、
  buildHourRows（#662）、buildStackedSeries/buildDirStackedSeries/buildWindowSummary/
  buildDirWindowSummary/buildDirTotals（堆叠/窗口/目录总量核心）+
  TrendMetric/TrendGranularity/TrendStackPart/TrendStackPoint/TrendWindowSummary
- src/trend/aggregator.ts：1329 → 701 行——保留状态容器（days/dirDays/hourDays/
  pending）与需要 this 的方法（apply/rebuild/rollupSnapshot/consume/pruneDays/
  查询薄壳）；查询方法改为委托纯函数（days/dirDays 显式传参）
- mutation-topology.json trend-aggregate 段 mutate 面并入两个新源文件（防新文件
  变异盲区），comment 更新；stryker:gen + stryker:check 通过
- gauntlet.config.json 段描述同步（scope/config/hardeningNote 文本；baseline*
  数字未动——重建动作见阶段报告）

设计决策与理由：
- 防「拆文件 = 共享 this」坏味道：所有迁出函数参数显式传入（days/dirDays/rows），
  不接触 this；凡需状态者（apply/rebuild/consume 等 IO 与状态操作）一律留在主类
- 模块间无循环依赖：aggregate-rows → types；aggregate-query → types/charts/
  aggregate-rows；aggregator → 三者（collector 侧仅 import type），构建与 typecheck
  实证无环
- 公开导出面兼容：metricValue/mergeAggRows/mergeDirRows/mergeHourRows/weekStartKey/
  lastNWeekKeys/lastNMonthKeys/monthRange/weekRange 与五类型经 aggregator.ts 尾部
  re-export——src/index.ts、src/report/*.ts 与 lib 产物对 "trend/aggregator.ts" 的
  import 路径不变（纯内部移动，lib 导出面抽查符号齐备）
- R8 台账守恒断言（unit-trend-ledger.test.ts）为本次拆分的顺序前置：拆分后守恒链
  漂移会第一时间红；test 全绿（21 pass）+ trend-aggregate 段变异实测 covered
  81.19%（threshold 60）验证行为等价
- 段无重叠验证：新文件仅属 trend-aggregate 段，位置去重口径（mutation-report-lib）
  不受影响；包级 baseline 数字待 observe 夜间重建

关联 issue：wingsky-1/dsh-plugin-hub#670（阶段三）
