+++ packages/dsh-provider-usage/src/trend/aggregator.ts
@@ -1,5 +1,5 @@
 /**
- * dsh-provider-usage/trend — 内存聚合（#503 M1）。
+ * dsh-provider-usage/trend — 内存聚合主类（#503 M1；D2 拆分后 #670 阶段三）。
  *
  * 两级聚合（方案定稿）：
  * - cells：day×provider(×model) 桶，apply 时实时累加（含未压实的历史日）；
@@ -19,13 +19,18 @@
  *
  * null 语义：桶内 token 求和 null-aware（无任何有效数字保持 null）；
  * 调用/轮次/工具计数独立累加，与 token 有无无关。
+ *
+ * D2 拆分（#670 阶段三）：本文件只保留状态容器（days/dirDays/hourDays/pending）与
+ * 需要访问状态的方法——apply/rebuild/rollupSnapshot/consume/prune 等 IO 与状态操作；
+ * 压实转换纯函数迁至 aggregate-rows.ts，查询投影纯函数迁至 aggregate-query.ts
+ * （二者均参数显式传入、不接触 this，防「拆文件 = 共享 this」坏味道；公开导出面
+ * 经本文件尾部 re-export 保持可达，src/index.ts 与 src/report/*.ts 的 import 路径不变）。
  */
 import { dayKey, lastNDayKeys } from "../charts.ts";
 import {
   sumToken,
   hourOfDay,
   TREND_ROW_VERSION,
-  TREND_UNIDENTIFIED,
   type TrendAggRow,
   type TrendCell,
   type TrendCounterRow,
@@ -35,56 +40,41 @@ import {
   type TrendTokens,
 } from "./types.ts";
 import type { TrendCallRecord, TrendCorrectRecord, TrendCounterRecord, TrendEmit } from "./collector.ts";
-
-/** 聚合指标（序列查询的取值维度；total = 四项 token 之和）。 */
-export type TrendMetric = "total" | "input" | "output" | "cacheRead" | "cacheWrite" | "calls";
-
-/** 序列粒度。 */
-export type TrendGranularity = "day" | "week" | "month";
-
-/** 堆叠柱单段（一个 provider 或 provider+model 组合在一个时间桶内的取值）。 */
-export interface TrendStackPart {
-  provider: string;
-  model: string | null;
-  value: number | null;
-}
-
-/** 堆叠柱单根（一个时间桶）。 */
-export interface TrendStackPoint {
-  /** 桶键：day=YYYY-MM-DD / week=周首日 / month=YYYY-MM。 */
-  key: string;
-  parts: TrendStackPart[];
-  /** 各段之和（null-aware；空桶 null）。 */
-  total: number | null;
-}
-
-/** 窗口摘要（趋势页汇总卡数据源）。 */
-export interface TrendWindowSummary {
-  /** 当前窗口指标总量。 */
-  total: number | null;
-  calls: number;
-  turns: number;
-  toolCalls: number;
-  /** 峰值桶键（total 最大的时间桶；无数据 null）。 */
-  peakKey: string | null;
-  /** 取值最大的适配器段。 */
-  top: { provider: string; model: string | null; value: number } | null;
-  /** 上一同等窗口指标总量（环比基准；无数据 null）。 */
-  prevTotal: number | null;
-  /** 上一窗口数据是否完整（起点早于数据起点 = false；false 时环比不可比，#503 M2.1）。 */
-  prevComplete: boolean;
-}
+import {
+  emptyCell,
+  emptyAggRow,
+  emptyDirRow,
+  emptyHourRow,
+  addDetailTo,
+  addCounterTo,
+  mergeCell,
+  sub,
+} from "./aggregate-rows.ts";
+import {
+  lastNWeekKeys,
+  lastNMonthKeys,
+  monthRange,
+  weekRange,
+  dayValueOf,
+  rangeValueOf,
+  buildDirRows,
+  buildHourRows,
+  buildStackedSeries,
+  buildDirStackedSeries,
+  buildWindowSummary,
+  buildDirWindowSummary,
+  buildDirTotals,
+  type TrendGranularity,
+  type TrendMetric,
+  type TrendStackPoint,
+  type TrendWindowSummary,
+} from "./aggregate-query.ts";
 
 /** 未压实行（内存持有；flush 时持久化，压实后移除）。 */
 export interface PendingEntry {
   row: TrendDetailRow | TrendCounterRow;
   /** 已 append 到当日分片（重启重建的行也为 true，防二次落盘）。 */
   persisted: boolean;
-}
-
-/** 单元格空壳。 */
-function emptyCell(): TrendCell {
-  return { input: null, output: null, cacheRead: null, cacheWrite: null, calls: 0, turns: 0, toolCalls: 0 };
 }
 
 export class TrendAggregator {
@@ -358,172 +348,18 @@ export class TrendAggregator {
 
   /**
    * 全量目录日桶快照（day 升序；#633 分片 b 报告快照与统计目录分布数据源）。
-   *
-   * 权威口径（复核 P1-1）：dirDays 单源快照——apply 平行累加 + rebuild 双分支读回
-   * 已覆盖全部 dir 事实，不再折算 pending 行（旧折算侧与 dirDays 并存时双算：
-   * 同事实重建 → 2×；互补事实 → 按键去重丢数；实测同日重启续 apply input 17 → 7）。
-   *
-   * 残差投影（本次修复，取代「重建时补造」）：目录面 = dirDays 快照 + 每日残差。
-   * 残差(day) = 该日聚合面（cells，全量事实）− 该日 dirDays 合计（有目录归属的事实），
-   * 非零则投影为一条 `{day, dir: TREND_UNIDENTIFIED}` 行——语义即「该日无目录信息的
-   * 数据」。这样两个查询面的日总量恒等（目录面 = 聚合面），且**不会双算**：
-   *
-   * - 旧分片（#633 之前的 agg 行，无 kind:"dir" 行）：cells 有值、dirDays 空 → 残差
-   *   = 全量 → 历史柱恢复且不丢数（此前 dir 面历史全 null，实测差 20 倍）；
-   * - 新分片（agg + dir 并存，同一批事实的两个投影）：cells 含 agg 行、dirDays 含 dir 行
-   *   → 残差 ≈ 0（同一事实相减相消）→ 不补造、不双算；故**禁止**在 rebuild 里对
-   *   agg 行补造未识别桶（会与 dir 行双算，反例：9-05 的 rjk2 calls=22 会变 44）；
-   * - 混版日（升级当天：旧 agg 行 + 新 dir 行并存）：残差 = 旧 agg 部分 → 归未识别，
-   *   新 dir 行照常分目录，既不丢升级前的历史、也不把新数据算进未识别；
-   * - 当日未压实：apply 对 cells 与 dirDays 平行累加 → 残差 ≈ 0。
-   *
-   * 不变量：∀day 目录面日合计 == 聚合面日合计（残差为负即数据不一致——目录行多于聚合行，
-   * 属双算/漂移征兆，投影按 0 处理并保持数值可解释，不产生负柱）。
-   * 残留边界（文档化口径）：残差只有日粒度（dirDays 无 provider 维度），故「无目录信息的
-   * 数据」只能整体归未识别桶，无法细分到 provider/model——目录面与 provider 面本互斥
-   * （见 routes/ui.ts 的 dir/byDir 分流），无消费方需要该交叉维度。
+   * 委托 buildDirRows——权威口径（dirDays 单源 + 每日残差归未识别）见 buildDirRows 注释。
    */
   dirRows(): TrendDirRow[] {
-    // 输出容器按 (day, dir) 唯一：dirDays 桶与残差行可能撞同一键（混版日已有
-    // (unidentified) 桶时），撞键即合并数值，绝不产出重复键行（公开面契约）。
-    // 保持 dirDays 分支的插入序（day 升序、day 内 dir 键插入序）——下游 dirTotals
-    // 用稳定排序，行序变化会改变「候选 calls 降序」的并列顺序。
-    const out: TrendDirRow[] = [];
-    const byKey = new Map<string, TrendDirRow>();
-    const put = (row: TrendDirRow): void => {
-      const key = `${row.day}\u0000${row.dir}`;
-      const cur = byKey.get(key);
-      if (cur === undefined) {
-        byKey.set(key, row);
-        out.push(row);
-        return;
-      }
-      cur.input = sumToken(cur.input, row.input);
-      cur.output = sumToken(cur.output, row.output);
-      cur.cacheRead = sumToken(cur.cacheRead, row.cacheRead);
-      cur.cacheWrite = sumToken(cur.cacheWrite, row.cacheWrite);
-      cur.calls += row.calls;
-      cur.turns += row.turns;
-      cur.toolCalls += row.toolCalls;
-    };
-    for (const day of [...this.dirDays.keys()].sort()) {
-      for (const [dir, cell] of this.dirDays.get(day)!) {
-        put({ v: TREND_ROW_VERSION, kind: "dir", day, dir, input: cell.input, output: cell.output, cacheRead: cell.cacheRead, cacheWrite: cell.cacheWrite, calls: cell.calls, turns: cell.turns, toolCalls: cell.toolCalls });
-      }
-    }
-    // 每日残差：聚合面（cells）− 目录面（dirDays）。cells 的键是 day → provider →
-    // model，逐层求和得该日全量；dirDays 的键是 day → dir，同法求该日目录合计。
-    const aggByDay = new Map<string, TrendCell>();
-    for (const [day, providers] of this.days) {
-      let cell = aggByDay.get(day);
-      if (cell === undefined) {
-        cell = emptyCell();
-        aggByDay.set(day, cell);
-      }
-      for (const models of providers.values()) {
-        for (const c of models.values()) {
-          cell.input = sumToken(cell.input, c.input);
-          cell.output = sumToken(cell.output, c.output);
-          cell.cacheRead = sumToken(cell.cacheRead, c.cacheRead);
-          cell.cacheWrite = sumToken(cell.cacheWrite, c.cacheWrite);
-          cell.calls += c.calls;
-          cell.turns += c.turns;
-          cell.toolCalls += c.toolCalls;
-        }
-      }
-    }
-    // 残差按 day 升序处理（cells 的 Map 键序是插入序，时钟回拨会让旧日新桶排在末尾；
-    // 公开方法契约声明 day 升序，故此处显式排序，不依赖插入序）。
-    for (const day of [...aggByDay.keys()].sort()) {
-      const cell = aggByDay.get(day)!;
-      let dirInput: number | null = null;
-      let dirOutput: number | null = null;
-      let dirCacheRead: number | null = null;
-      let dirCacheWrite: number | null = null;
-      let dirCalls = 0;
-      let dirTurns = 0;
-      let dirToolCalls = 0;
-      const dirs = this.dirDays.get(day);
-      if (dirs !== undefined) {
-        for (const c of dirs.values()) {
-          dirInput = sumToken(dirInput, c.input);
-          dirOutput = sumToken(dirOutput, c.output);
-          dirCacheRead = sumToken(dirCacheRead, c.cacheRead);
-          dirCacheWrite = sumToken(dirCacheWrite, c.cacheWrite);
-          dirCalls += c.calls;
-          dirTurns += c.turns;
-          dirToolCalls += c.toolCalls;
-        }
-      }
-      // 残差 = 聚合面 − 目录面（null-aware：双方皆 null → 0/无残差；单侧 null 按 0 计）
-      const input = diffToken(cell.input, dirInput);
-      const output = diffToken(cell.output, dirOutput);
-      const cacheRead = diffToken(cell.cacheRead, dirCacheRead);
-      const cacheWrite = diffToken(cell.cacheWrite, dirCacheWrite);
-      const calls = cell.calls - dirCalls;
-      const turns = cell.turns - dirTurns;
-      const toolCalls = cell.toolCalls - dirToolCalls;
-      // 全部为 0/空 = 该日目录面已覆盖全量（新分片常态）→ 不补造行。
-      // 负残差（目录面多于聚合面）同样走此分支（不产行）：属数据不一致征兆
-      // （例如明细目录误落 dir 行——已由 readDetailShard 白名单阻断），此时目录面
-      // 日合计会大于聚合面，README「总量守恒」节已注明该边界不保证恒等。
-      const hasResidual =
-        calls > 0 || turns > 0 || toolCalls > 0 || (input ?? 0) > 0 || (output ?? 0) > 0 || (cacheRead ?? 0) > 0 || (cacheWrite ?? 0) > 0;
-      if (!hasResidual) continue;
-      // 同键（该日已有 (unidentified) 桶，混版日常态）由 put 合并到既有行，
-      // 保证 (day, dir) 键唯一；负值按 0 处理（不产生负柱）。
-      put({
-        v: TREND_ROW_VERSION,
-        kind: "dir",
-        day,
-        dir: TREND_UNIDENTIFIED,
-        input: input !== null && input > 0 ? input : null,
-        output: output !== null && output > 0 ? output : null,
-        cacheRead: cacheRead !== null && cacheRead > 0 ? cacheRead : null,
-        cacheWrite: cacheWrite !== null && cacheWrite > 0 ? cacheWrite : null,
-        calls: calls > 0 ? calls : 0,
-        turns: turns > 0 ? turns : 0,
-        toolCalls: toolCalls > 0 ? toolCalls : 0,
-      });
-    }
-    // 顺序契约：dirDays 分支按 day 升序、day 内按 dir 键插入序输出；残差行按 day
-    // 升序追加在尾段（已显式排序，不依赖 cells 插入序——时钟回拨会让旧日新桶排在
-    // 末尾）。**不做全局 sort(day, dir)**：它会把残差行的未识别键排到该日首位
-    // （"(" 字典序最前），改变 dirTotals 的插入序；dirTotals 用稳定排序（同 calls
-    // 保持插入序），下游「候选 calls 降序」并列顺序因此会被打破。
-    return out;
+    return buildDirRows(this.days, this.dirDays);
   }
 
   /**
-   * 全量小时日桶快照（day 升序、hour 升序；#662 报告快照 byHour/byPeriod/peakHour/
-   * coveredDays 数据源）。
-   * 权威口径（对齐 dirRows 单源约定）：hourDays 单源快照——apply 平行累加 + rebuild
-   * 双分支读回已覆盖全部 hour 事实，不折算 pending（防双算）；**不做残差投影**
-   * （detail/counter 必有 time、无缺键事实；旧分片缺 hour 行是物理缺失，报告侧
-   * coveredDays 守卫负责降级，不投影补造伪事实）。
+   * 全量小时日桶快照（day 升序、hour 升序；#662 报告快照数据源）。
+   * 委托 buildHourRows——hourDays 单源快照、不做残差投影（口径见 buildHourRows 注释）。
    */
   hourRows(): TrendHourRow[] {
-    const out: TrendHourRow[] = [];
-    for (const day of [...this.hourDays.keys()].sort()) {
-      const byHour = this.hourDays.get(day)!;
-      for (const hour of [...byHour.keys()].sort((a, b) => a - b)) {
-        const cell = byHour.get(hour)!;
-        out.push({
-          v: TREND_ROW_VERSION,
-          kind: "hour",
-          day,
-          hour,
-          input: cell.input,
-          output: cell.output,
-          cacheRead: cell.cacheRead,
-          cacheWrite: cell.cacheWrite,
-          calls: cell.calls,
-          turns: cell.turns,
-          toolCalls: cell.toolCalls,
-        });
-      }
-    }
-    return out;
+    return buildHourRows(this.hourDays);
   }
 
   /**
@@ -538,33 +374,7 @@ export class TrendAggregator {
     dir: string | undefined,
     now: number,
   ): { series: TrendStackPoint[]; dirs: Array<{ dir: string }> } {
-    const keys = this.granKeys(n, gran, now);
-    const ranges = new Map(keys.map((k) => [k, this.granRange(k, gran)] as const));
-    // 复核 P1-3：行快照提出桶循环（原实现每桶 this.dirRows() 全量快照，O(桶×行)
-    // 单请求重复；对齐 seriesStacked 的 P1-2 先例——range 一次算全 + 快照单次
-    // 取用，桶循环内只做窗口过滤消费）。
-    const rows = this.dirRows();
-    const series: TrendStackPoint[] = keys.map((key) => {
-      const partsMap = new Map<string, TrendStackPart>();
-      const range = ranges.get(key)!;
-      for (const row of rows) {
-        if (row.day < range.start || row.day > range.end) continue;
-        if (dir !== undefined && row.dir !== dir) continue;
-        const v = metricValue(row, metric);
-        const cur = partsMap.get(row.dir);
-        if (cur !== undefined) cur.value = sumToken(cur.value, v);
-        else partsMap.set(row.dir, { provider: row.dir, model: null, value: v });
-      }
-      const parts = [...partsMap.values()];
-      let total: number | null = null;
-      for (const pt of parts) total = sumToken(total, pt.value);
-      return { key, parts, total };
-    });
-    const legend = new Set<string>();
-    for (const point of series) {
-      for (const pt of point.parts) legend.add(pt.provider);
-    }
-    return { series, dirs: [...legend].sort().map((d) => ({ dir: d })) };
+    return buildDirStackedSeries(this.dirRows(), n, gran, metric, dir, now);
   }
 
   /**
@@ -581,64 +391,7 @@ export class TrendAggregator {
     now: number,
     dirSeries?: TrendStackPoint[],
   ): TrendWindowSummary {
-    const keys = this.granKeys(n, gran, now);
-    const curRange = {
-      start: this.granRange(keys[0], gran).start,
-      end: this.granRange(keys[keys.length - 1], gran).end,
-    };
-    const prevKeys = this.granKeys(n, gran, this.prevWindowAnchor(now, n, gran));
-    const prevRange = {
-      start: this.granRange(prevKeys[0], gran).start,
-      end: this.granRange(prevKeys[prevKeys.length - 1], gran).end,
-    };
-    let total: number | null = null;
-    let calls = 0;
-    let turns = 0;
-    let toolCalls = 0;
-    let prevTotal: number | null = null;
-    let peakKey: string | null = null;
-    let peakVal = -1;
-    let top: { provider: string; model: string | null; value: number } | null = null;
-    const series = dirSeries ?? this.dirStacked(n, gran, metric, dir, now).series;
-    for (const point of series) {
-      if (point.total !== null && point.total > peakVal) {
-        peakVal = point.total;
-        peakKey = point.key;
-      }
-      for (const pt of point.parts) {
-        if (pt.value !== null && (top === null || pt.value > top.value)) {
-          top = { provider: pt.provider, model: pt.model, value: pt.value };
-        }
-      }
-    }
-    const rows = this.dirRows();
-    let firstDirDay: string | null = null;
-    // 复核 P1-3：单遍遍历（原实现 3 次全量遍历 rows——最早日 + 当前窗口 + 上一
-    // 窗口各一遍；窗口区间互斥，同遍累加语义不变）。
-    for (const row of rows) {
-      if (firstDirDay === null || row.day < firstDirDay) firstDirDay = row.day;
-      const inDir = dir === undefined || row.dir === dir;
-      if (inDir && row.day >= curRange.start && row.day <= curRange.end) {
-        total = sumToken(total, metricValue(row, metric));
-        calls += row.calls;
-        turns += row.turns;
-        toolCalls += row.toolCalls;
-      }
-      if (inDir && row.day >= prevRange.start && row.day <= prevRange.end) {
-        prevTotal = sumToken(prevTotal, metricValue(row, metric));
-      }
-    }
-    return {
-      total,
-      calls,
-      turns,
-      toolCalls,
-      peakKey,
-      top,
-      prevTotal,
-      // 与 windowSummary.prevComplete 同口径：prev 窗口起点早于目录数据起点 → 环比不可比
-      prevComplete: prevRange.start >= (firstDirDay ?? "9999-12-31"),
-    };
+    return buildDirWindowSummary(this.dirRows(), n, gran, metric, dir, now, dirSeries);
   }
 
   /**
@@ -650,18 +403,7 @@ export class TrendAggregator {
     endDay: string,
     metric: TrendMetric = "total",
   ): Array<{ dir: string; calls: number; total: number | null }> {
-    const byDir = new Map<string, { dir: string; calls: number; total: number | null }>();
-    for (const row of this.dirRows()) {
-      if (row.day < startDay || row.day > endDay) continue;
-      let cur = byDir.get(row.dir);
-      if (cur === undefined) {
-        cur = { dir: row.dir, calls: 0, total: null };
-        byDir.set(row.dir, cur);
-      }
-      cur.calls += row.calls;
-      cur.total = sumToken(cur.total, metricValue(row, metric));
-    }
-    return [...byDir.values()].sort((a, b) => b.calls - a.calls);
+    return buildDirTotals(this.dirRows(), startDay, endDay, metric);
   }
 
   /**
@@ -829,24 +571,25 @@ export class TrendAggregator {
 
   /** 近 n 日序列（含今日；空日补 null——零 usage 语义，非 0）。 */
   seriesDays(n: number, now: number, metric: TrendMetric, provider?: string): Array<{ day: string; value: number | null }> {
-    return lastNDayKeys(n, now).map((day) => ({ day, value: this.dayValue(day, metric, provider) }));
+    return lastNDayKeys(n, now).map((day) => ({ day, value: dayValueOf(this.days, day, metric, provider) }));
   }
 
   /** 近 n 周序列（周一起点；key = 周首日 day key）。 */
   seriesWeeks(n: number, now: number, metric: TrendMetric, provider?: string): Array<{ day: string; value: number | null }> {
     const keys = lastNWeekKeys(n, now);
-    return keys.map((day) => ({ day, value: this.rangeValue(weekRange(day), metric, provider) }));
+    return keys.map((day) => ({ day, value: rangeValueOf(this.days, weekRange(day), metric, provider) }));
   }
 
   /** 近 n 月序列（key = YYYY-MM）。 */
   seriesMonths(n: number, now: number, metric: TrendMetric, provider?: string): Array<{ day: string; value: number | null }> {
     const keys = lastNMonthKeys(n, now);
-    return keys.map((key) => ({ day: key, value: this.rangeValue(monthRange(key), metric, provider) }));
+    return keys.map((key) => ({ day: key, value: rangeValueOf(this.days, monthRange(key), metric, provider) }));
   }
 
   /**
    * 堆叠柱序列（/trend 路由数据源）：每时间桶按 provider（byModel 时细到
    * provider+model）拆段；providers 为图例并集（窗口内出现过的段）。
+   * 委托 buildStackedSeries（days 显式传参，查询投影纯计算）。
    */
   seriesStacked(
     n: number,
@@ -856,46 +599,13 @@ export class TrendAggregator {
     byModel: boolean,
     now: number,
   ): { series: TrendStackPoint[]; providers: Array<{ provider: string; model: string | null }> } {
-    const keys = this.granKeys(n, gran, now);
-    // 桶日区间一次算全（评审 P1-2：原实现对每桶在 days 日循环内重复调 granRange）
-    const ranges = new Map(keys.map((k) => [k, this.granRange(k, gran)] as const));
-    const series: TrendStackPoint[] = keys.map((key) => {
-      const partsMap = new Map<string, TrendStackPart>();
-      const range = ranges.get(key)!;
-      for (const [day, models] of this.days) {
-        if (day < range.start || day > range.end) continue;
-        for (const [p, cells] of models) {
-          if (provider !== undefined && p !== provider) continue;
-          for (const [m, cell] of cells) {
-            const id = byModel ? `${p}\u0000${m ?? ""}` : p;
-            const v = metricValue(cell, metric);
-            const cur = partsMap.get(id);
-            if (cur !== undefined) cur.value = sumToken(cur.value, v);
-            else partsMap.set(id, { provider: p, model: byModel ? m : null, value: v });
-          }
-        }
-      }
-      const parts = [...partsMap.values()];
-      let total: number | null = null;
-      for (const pt of parts) total = sumToken(total, pt.value);
-      return { key, parts, total };
-    });
-    const legend = new Map<string, { provider: string; model: string | null }>();
-    for (const point of series) {
-      for (const pt of point.parts) {
-        legend.set(byModel ? `${pt.provider}\u0000${pt.model ?? ""}` : pt.provider, {
-          provider: pt.provider,
-          model: byModel ? pt.model : null,
-        });
-      }
-    }
-    return { series, providers: [...legend.values()] };
+    return buildStackedSeries(this.days, n, gran, metric, provider, byModel, now);
   }
 
   /**
    * 窗口摘要（当前 n 桶 + 上一同等窗口环比基准）。
-   * @param stackSeries 可选传入路由已算好的堆叠序列（n/gran/metric/provider 必须与本
-   *   调用一致）——复用峰值/Top 遍历，消除单请求双算（评审 P1-2）；缺省时内部自算。
+   * 委托 buildWindowSummary；stackSeries 可选传入路由已算好的堆叠序列复用遍历
+   * （须与 n/gran/metric/provider 同参，见 aggregator.windowSummary 契约）。
    */
   windowSummary(
     n: number,
@@ -905,120 +615,7 @@ export class TrendAggregator {
     now: number,
     stackSeries?: TrendStackPoint[],
   ): TrendWindowSummary {
-    const keys = this.granKeys(n, gran, now);
-    const curRange = {
-      start: this.granRange(keys[0], gran).start,
-      end: this.granRange(keys[keys.length - 1], gran).end,
-    };
-    // 上一窗口：同长键序列整体前移 n 桶（day/week/month 各自步进语义）
-    const prevKeys = this.granKeys(n, gran, this.prevWindowAnchor(now, n, gran));
-    const prevRange = {
-      start: this.granRange(prevKeys[0], gran).start,
-      end: this.granRange(prevKeys[prevKeys.length - 1], gran).end,
-    };
-    let total: number | null = null;
-    let calls = 0;
-    let turns = 0;
-    let toolCalls = 0;
-    let peakKey: string | null = null;
-    let peakVal = -1;
-    let top: { provider: string; model: string | null; value: number } | null = null;
-    // 复用路由已算的 stack 序列（评审 P1-2：消除单请求双算；缺省自算保持独立可用）
-    const series = stackSeries ?? this.seriesStacked(n, gran, metric, provider, false, now).series;
-    for (const point of series) {
-      if (point.total !== null && point.total > peakVal) {
-        peakVal = point.total;
-        peakKey = point.key;
-      }
-      for (const pt of point.parts) {
-        if (pt.value !== null && (top === null || pt.value > top.value)) {
-          top = { provider: pt.provider, model: pt.model, value: pt.value };
-        }
-      }
-    }
-    for (const { cell } of this.rangeCells(curRange, provider)) {
-      total = sumToken(total, metricValue(cell, metric));
-      calls += cell.calls;
-      turns += cell.turns;
-      toolCalls += cell.toolCalls;
-    }
-    return {
-      total,
-      calls,
-      turns,
-      toolCalls,
-      peakKey,
-      top,
-      prevTotal: this.rangeValue(prevRange, metric, provider),
-      // 上一窗口起点早于数据起点（内存最早日，即留存/起算边缘）→ 基准不完整，环比不可比（#503 M2.1）
-      prevComplete: prevRange.start >= (firstDayKeyOf(this.days) ?? "9999-12-31"),
-    };
-  }
-
-  // ---------------------------------------------------------------- 内部
-
-  /** 粒度桶键序列（升序，含当前桶）。 */
-  private granKeys(n: number, gran: TrendGranularity, now: number): string[] {
-    if (gran === "week") return lastNWeekKeys(n, now);
-    if (gran === "month") return lastNMonthKeys(n, now);
-    return lastNDayKeys(n, now);
-  }
-
-  /** 桶键 → 本地日区间。 */
-  private granRange(key: string, gran: TrendGranularity): { start: string; end: string } {
-    if (gran === "week") return weekRange(key);
-    if (gran === "month") return monthRange(key);
-    return { start: key, end: key };
-  }
-
-  /** 上一窗口锚点（把 now 前移 n 桶，得到 prevKeys 与当前窗口不重叠的时点）。 */
-  private prevWindowAnchor(now: number, n: number, gran: TrendGranularity): number {
-    const d = new Date(now);
-    if (gran === "month") {
-      d.setMonth(d.getMonth() - n);
-      return d.getTime();
-    }
-    const days = gran === "week" ? n * 7 : n;
-    d.setDate(d.getDate() - days);
-    return d.getTime();
-  }
-
-  /** 日区间内的全部 (provider, model, cell)（升序日；provider 过滤可选）。 */
-  private rangeCells(range: { start: string; end: string }, provider?: string): Array<{ provider: string; model: string | null; cell: TrendCell }> {
-    const out: Array<{ provider: string; model: string | null; cell: TrendCell }> = [];
-    for (const [day, models] of this.days) {
-      if (day < range.start || day > range.end) continue;
-      for (const [p, cells] of models) {
-        if (provider !== undefined && p !== provider) continue;
-        for (const [m, cell] of cells) out.push({ provider: p, model: m, cell });
-      }
-    }
-    return out;
-  }
-
-  /** 单日取值（provider 过滤可选；跨 model 求和）。 */
-  private dayValue(day: string, metric: TrendMetric, provider?: string): number | null {
-    const models = this.days.get(day);
-    if (models === undefined) return null;
-    let acc: number | null = null;
-    for (const [p, cells] of models) {
-      if (provider !== undefined && p !== provider) continue;
-      for (const cell of cells.values()) acc = sumToken(acc, metricValue(cell, metric));
-    }
-    return acc;
-  }
-
-  /** 日区间取值（[startDay, endDay] 闭区间，按本地日字典序比较）。 */
-  private rangeValue(range: { start: string; end: string }, metric: TrendMetric, provider?: string): number | null {
-    let acc: number | null = null;
-    for (const [day, models] of this.days) {
-      if (day < range.start || day > range.end) continue;
-      for (const [p, cells] of models) {
-        if (provider !== undefined && p !== provider) continue;
-        for (const cell of cells.values()) acc = sumToken(acc, metricValue(cell, metric));
-      }
-    }
-    return acc;
+    return buildWindowSummary(this.days, n, gran, metric, provider, now, stackSeries);
   }
 
   /** 统计摘要（health 观测面）。 */
@@ -1082,248 +679,24 @@ export class TrendAggregator {
   }
 }
 
-// ---------------------------------------------------------------- 纯函数
+// ---------------------------------------------------------------- 纯函数 re-export（公开面兼容）
 
-/** token 差（cell 回退用）：有值→null 时增量为 -old（cell 同步扣减，保证内存聚合
- *  与校正后的落盘明细一致）；双方皆 null 增量 0（null-aware 无变化）。 */
-function sub(oldV: number | null, newV: number | null): number | null {
-  if (newV === null) return oldV === null ? null : -oldV;
-  if (oldV === null) return newV;
-  return newV - oldV;
-}
-
-/**
- * 残差投影的 token 差（#633 修复）：聚合面 − 目录面，null-aware。
- * 双方皆 null → null（无该维度事实）；单侧 null 按 0 参与（另一侧有值即有残差）；
- * 负值保留给调用方判定（调用方按 0 处理并依赖「日总量恒等」断言暴露不一致）。
- */
-function diffToken(aggV: number | null, dirV: number | null): number | null {
-  if (aggV === null && dirV === null) return null;
-  return (aggV ?? 0) - (dirV ?? 0);
-}
-
-/** 内存日桶的最早 day key（无数据返回 null；day key 字典序即时间序）。 */
-function firstDayKeyOf(days: Map<string, unknown>): string | null {
-  let first: string | null = null;
-  for (const day of days.keys()) {
-    if (first === null || day < first) first = day;
-  }
-  return first;
-}
-
-/** 空聚合行（压实折算起点；字段与 mergeCell 消费的 agg 行同构）。 */
-function emptyAggRow(day: string, provider: string, model: string | null): TrendAggRow {
-  return {
-    v: TREND_ROW_VERSION,
-    kind: "agg",
-    day,
-    provider,
-    model,
-    input: null,
-    output: null,
-    cacheRead: null,
-    cacheWrite: null,
-    calls: 0,
-    turns: 0,
-    toolCalls: 0,
-  };
-}
-
-/** 空目录汇总行（#633 A4；与 agg 行十数值字段同构，仅键换成 dir）。 */
-function emptyDirRow(day: string, dir: string): TrendDirRow {
-  return {
-    v: TREND_ROW_VERSION,
-    kind: "dir",
-    day,
-    dir,
-    input: null,
-    output: null,
-    cacheRead: null,
-    cacheWrite: null,
-    calls: 0,
-    turns: 0,
-    toolCalls: 0,
-  };
-}
-
-/** 空小时汇总行（#662；与 agg/dir 行同构，键换成 hour；「落盘即定型」的产出起点）。 */
-function emptyHourRow(day: string, hour: number): TrendHourRow {
-  return {
-    v: TREND_ROW_VERSION,
-    kind: "hour",
-    day,
-    hour,
-    input: null,
-    output: null,
-    cacheRead: null,
-    cacheWrite: null,
-    calls: 0,
-    turns: 0,
-    toolCalls: 0,
-  };
-}
-
-/** 明细行并入聚合/目录/小时汇总行（三者字段同构，同函数复用；null-aware 求和）。 */
-function addDetailTo(agg: TrendAggRow | TrendDirRow | TrendHourRow, row: TrendDetailRow): void {
-  agg.calls += 1;
-  agg.input = sumToken(agg.input, row.input);
-  agg.output = sumToken(agg.output, row.output);
-  agg.cacheRead = sumToken(agg.cacheRead, row.cacheRead);
-  agg.cacheWrite = sumToken(agg.cacheWrite, row.cacheWrite);
-}
-
-/** 计数行并入聚合/目录/小时汇总行（同上，同函数复用）。 */
-function addCounterTo(agg: TrendAggRow | TrendDirRow | TrendHourRow, turns: number, toolCalls: number): void {
-  agg.turns += turns;
-  agg.toolCalls += toolCalls;
-}
-
-/** 聚合行并入 cell（重建用；null-aware。agg/dir/hour 汇总行十数值字段同构，同函数复用）。 */
-function mergeCell(cell: TrendCell, row: TrendAggRow | TrendDirRow | TrendHourRow): void {
-  cell.calls += row.calls;
-  cell.turns += row.turns;
-  cell.toolCalls += row.toolCalls;
-  cell.input = sumToken(cell.input, row.input);
-  cell.output = sumToken(cell.output, row.output);
-  cell.cacheRead = sumToken(cell.cacheRead, row.cacheRead);
-  cell.cacheWrite = sumToken(cell.cacheWrite, row.cacheWrite);
-}
-
-/** cell 的指标取值（total = 四项 token 之和）。 */
-export function metricValue(cell: TrendCell, metric: TrendMetric): number | null {
-  switch (metric) {
-    case "total":
-      return sumToken(sumToken(cell.input, cell.output), sumToken(cell.cacheRead, cell.cacheWrite));
-    case "input":
-      return cell.input;
-    case "output":
-      return cell.output;
-    case "cacheRead":
-      return cell.cacheRead;
-    case "cacheWrite":
-      return cell.cacheWrite;
-    case "calls":
-      return cell.calls;
-  }
-}
-
-/**
- * 聚合行合并（flush 压实写盘前与既有聚合分片合并——迟到旧日行场景防覆盖丢数；
- * null-aware 求和，同 (provider, model) 键累加）。
- */
-export function mergeAggRows(base: TrendAggRow[], add: TrendAggRow[]): TrendAggRow[] {
-  const byKey = new Map<string, TrendAggRow>();
-  for (const r of [...base, ...add]) {
-    const key = `${r.provider}\u0000${r.model ?? ""}`;
-    const cur = byKey.get(key);
-    if (cur === undefined) {
-      byKey.set(key, { ...r });
-      continue;
-    }
-    cur.calls += r.calls;
-    cur.turns += r.turns;
-    cur.toolCalls += r.toolCalls;
-    cur.input = sumToken(cur.input, r.input);
-    cur.output = sumToken(cur.output, r.output);
-    cur.cacheRead = sumToken(cur.cacheRead, r.cacheRead);
-    cur.cacheWrite = sumToken(cur.cacheWrite, r.cacheWrite);
-  }
-  return [...byKey.values()];
-}
-
-/**
- * dir 汇总行合并（#633 A4 flush 压实写盘前与既有聚合分片内的 dir 行合并——
- * 迟到旧日行场景防覆盖丢数；十数值字段与 mergeAggRows 完全同构，null-aware
- * 求和，同 dir 键累加。输出保持输入相对顺序：base 在前（dir 行位于分片尾段））。
- */
-export function mergeDirRows(base: TrendDirRow[], add: TrendDirRow[]): TrendDirRow[] {
-  const byKey = new Map<string, TrendDirRow>();
-  for (const r of [...base, ...add]) {
-    const cur = byKey.get(r.dir);
-    if (cur === undefined) {
-      byKey.set(r.dir, { ...r });
-      continue;
-    }
-    cur.calls += r.calls;
-    cur.turns += r.turns;
-    cur.toolCalls += r.toolCalls;
-    cur.input = sumToken(cur.input, r.input);
-    cur.output = sumToken(cur.output, r.output);
-    cur.cacheRead = sumToken(cur.cacheRead, r.cacheRead);
-    cur.cacheWrite = sumToken(cur.cacheWrite, r.cacheWrite);
-  }
-  return [...byKey.values()];
-}
-
-/**
- * hour 汇总行合并（#662，flush 压实写盘前与既有聚合分片内的 hour 行合并——
- * 迟到旧日行二次压实防丢防重；与 mergeDirRows 完全同构，同 hour 键累加，
- * 输出保持输入相对顺序：base 在前（hour 行位于分片尾段））。
- */
-export function mergeHourRows(base: TrendHourRow[], add: TrendHourRow[]): TrendHourRow[] {
-  const byKey = new Map<number, TrendHourRow>();
-  for (const r of [...base, ...add]) {
-    const cur = byKey.get(r.hour);
-    if (cur === undefined) {
-      byKey.set(r.hour, { ...r });
-      continue;
-    }
-    cur.calls += r.calls;
-    cur.turns += r.turns;
-    cur.toolCalls += r.toolCalls;
-    cur.input = sumToken(cur.input, r.input);
-    cur.output = sumToken(cur.output, r.output);
-    cur.cacheRead = sumToken(cur.cacheRead, r.cacheRead);
-    cur.cacheWrite = sumToken(cur.cacheWrite, r.cacheWrite);
-  }
-  return [...byKey.values()];
-}
-
-/** 周一为起点的周首日 day key（本地时区；DST 安全——逐日回退不用毫秒减法）。 */
-export function weekStartKey(t: number): string {
-  const d = new Date(t);
-  const dow = d.getDay(); // 0=周日
-  const back = (dow + 6) % 7; // 距周一的天数
-  d.setDate(d.getDate() - back);
-  return dayKey(d.getTime());
-}
-
-/** 近 n 个周首日 key（升序，含当前周）。 */
-export function lastNWeekKeys(n: number, now: number): string[] {
-  const d = new Date(now);
-  d.setDate(d.getDate() - ((d.getDay() + 6) % 7));
-  const keys: string[] = [];
-  for (let i = 0; i < n; i += 1) {
-    keys.unshift(dayKey(d.getTime()));
-    d.setDate(d.getDate() - 7);
-  }
-  return keys;
-}
-
-/** 近 n 个月 key（YYYY-MM，升序，含当月；本地时区逐月回退）。 */
-export function lastNMonthKeys(n: number, now: number): string[] {
-  const d = new Date(now);
-  const keys: string[] = [];
-  for (let i = 0; i < n; i += 1) {
-    keys.unshift(`${d.getFullYear()}-${String(d.getMonth() + 1).padStart(2, "0")}`);
-    d.setDate(1);
-    d.setMonth(d.getMonth() - 1);
-  }
-  return keys;
-}
-
-/** 月 key 的日区间（本地日字典序；月末取该月最后一天，DST 安全）。 */
-export function monthRange(key: string): { start: string; end: string } {
-  const [y, m] = key.split("-").map(Number);
-  const start = new Date(y, m - 1, 1);
-  const end = new Date(y, m, 0); // 次月 0 日 = 本月末日
-  return { start: dayKey(start.getTime()), end: dayKey(end.getTime()) };
-}
-
-/** 周首日 key 的日区间（7 天）。 */
-export function weekRange(weekStart: string): { start: string; end: string } {
-  const [y, m, d] = weekStart.split("-").map(Number);
-  const start = new Date(y, m - 1, d);
-  const endDate = new Date(y, m - 1, d + 6);
-  return { start: dayKey(start.getTime()), end: dayKey(endDate.getTime()) };
-}
+// 以下符号原定义于本文件，D2 拆分迁至 aggregate-rows.ts / aggregate-query.ts 后经此处
+// re-export——src/index.ts、src/report/*.ts 与 lib 产物对 "trend/aggregator.ts" 的
+// import 路径保持不变（纯内部移动，公开导出面不破坏）。
+export {
+  metricValue,
+  weekStartKey,
+  lastNWeekKeys,
+  lastNMonthKeys,
+  monthRange,
+  weekRange,
+} from "./aggregate-query.ts";
+export { mergeAggRows, mergeDirRows, mergeHourRows } from "./aggregate-rows.ts";
+export type {
+  TrendGranularity,
+  TrendMetric,
+  TrendStackPart,
+  TrendStackPoint,
+  TrendWindowSummary,
+} from "./aggregate-query.ts";
