+++ scripts/data/mutation-topology.json
@@ -336,6 +336,8 @@
         "trend-aggregate": {
           "mutate": [
             "packages/dsh-provider-usage/src/trend/aggregator.ts",
+            "packages/dsh-provider-usage/src/trend/aggregate-rows.ts",
+            "packages/dsh-provider-usage/src/trend/aggregate-query.ts",
             "packages/dsh-provider-usage/src/trend/store.ts",
             "packages/dsh-provider-usage/src/trend/index.ts"
           ],
@@ -343,7 +345,7 @@
             "!packages/dsh-provider-usage/src/client/**",
             "!packages/dsh-provider-usage/src/types.ts"
           ],
-          "comment": "阶段零新段（D17 分层变异）：E2 聚合/压实/存储面——aggregator+store+index（目录维度与 provider 面查询）；testFiles 含 unit-trend/unit-trend-view（杀灭面为该层测试）；aggregator 单文件 5-8 分钟，独立段防超时。"
+          "comment": "阶段零新段（D17 分层变异）：E2 聚合/压实/存储面——aggregator+store+index；阶段三（#670 D2）aggregator 拆分为 aggregate-rows（压实转换纯函数）+ aggregate-query（查询投影纯函数），两新文件并入 mutate 面（防新文件变异盲区）；testFiles 含 unit-trend/unit-trend-ledger/unit-trend-view（杀灭面为该层测试，R8 台账守恒断言覆盖守恒链）；aggregator 单文件 5-8 分钟，独立段防超时。"
         },
         "trend-collect": {
           "mutate": [
